Astraphobia Trigger Option
Astraphobia is the fear of thunderstorms, I have lilapsophobia which is the phobia of tornadoes and sometimes can be seen as a fear of severe thunderstorms as well as tornadoes and hurricanes. I am adding this feature in hopes that it will assist some people who have very mild triggers from Week 2's occasional blasts and near realistic thunder sounds that cause me skin to crawl ever so slightly, not hindering my gameplay but I know that other's aren't as fortunate.