Fixed Health Bar color functionality
## What's wrong with the heath bar color functionality?
You may have noticed that changing the player in debug does not change the player health bar color (i.e changing to say pico still has the blue color on the health bar). This does work with the opponents though, as it changes based on the opponent.

## What I did
I created two variables colorOne and colorTwo as Ints, and where the health bar is drawn, I made it 1st check the player, and changes colorOne to what character is being played (being bf makes it blue, changing to say dad will turn it purple). Then, it checks the opponent and changes colorTwo to the right color (if senpai was the opponent, the other side would be peach, or being pico makes it green).